chore: change copy function

+++ src/main.rs
@@ -1,9 +1,9 @@
 use chrono::prelude::*;
 use clap::Parser;
 use notify::{RecommendedWatcher, RecursiveMode, Result, Watcher};
-use std::fs::File;
 use std::io::{self, Read, Write};
 use std::path::Path;
+use std::process::exit;
 use std::thread;
 use std::time::Duration;
 
@@ -18,9 +18,6 @@ fn main() -> Result<()> {
     let args = Args::parse();
     let file_path = args.input;
     let output_file = args.output;
-    let path = String::from(file_path.clone());
-
-    let new_path = file_path.clone();
 
     let copy_source_path = Path::new(&file_path).to_path_buf();
     let source_path = Path::new(&file_path);
@@ -40,35 +37,35 @@ fn main() -> Result<()> {
                 &output_file,
                 local.format("%Y-%m-%d %H:%M:%S")
             );
-            let new_source_path = Path::new(&new_path);
-            println!("{}", new_source_path.exists());
             copy_file(&copy_source_path, &copy_target_path);
+            // copy_file(&source_path_before_listening.clone(), &copy_target_path);
             println!("updated!");
         }
         Err(e) => println!("watch error: {:?}", e),
     })?;
 
     loop {
-        if source_path.exists() {
+        if source_path.exists() && source_path.is_file() {
+            println!("find file");
             copy_file(&source_path_before_listening, &target_path_before_listening);
-            watcher.watch(
-                Path::new(String::from(path.clone()).as_str()),
-                RecursiveMode::Recursive,
-            )?;
+            watcher.watch(source_path, RecursiveMode::NonRecursive)?;
             break;
+        } else if source_path.is_dir() {
+            println!("input can only be a file. not a dir.");
+            exit(0x0100);
         } else {
-            thread::sleep(Duration::from_secs(1));
+            println!("no file");
         }
+        thread::sleep(Duration::from_secs(1));
     }
 
-    loop {}
+    loop {
+        thread::sleep(Duration::from_millis(100));
+    }
     Ok(())
 }
 
 fn copy_file(source: &Path, target: &Path) {
-    println!("import ");
-    println!("{:?}", source);
-    let mut src_file = File::open(source).expect("");
-    let mut dst_file = File::create(target).expect("");
-    io::copy(&mut src_file, &mut dst_file);
+    println!("start copy: {:?}", source);
+    std::fs::copy(source, target);
 }
